// src/types/response.ts
interface responseData {
  cmuitaccount_name: string;
  cmuitaccount: string;
  student_id: string;
  firstname_TH: string;
  lastname_TH: string;
  organization_name_TH: string;
  organization_name_EN: string;
  organization_code: string;
  itaccounttype_EN: string;
  itaccounttype_TH: string;
}

export default responseData;